requirements.txt
editfrontmatter==0.0.1
python_frontmatter==1.0.0
